+++ tests/test_concurrency_async.py
@@ -61,7 +61,10 @@ async def test_concurrent_execution(aconn_cls, dsn):
 async def canceller(aconn, errors):
     try:
         await asyncio.sleep(0.5)
-        aconn.cancel()
+        try:
+            await aconn.cancel_safe()
+        except e.NotSupportedError:
+            aconn.cancel()
     except Exception as exc:
         errors.append(exc)